added validator to check for non-existing URLs
I added a new validator to oscar.core.validators that validates
absolute and relative URL and checks if the URL already exists. We
use this functionality when creating flatpages in the dashboard to
check if the URL can be used.
The validation includes *two* calls to ExtendedURLValidator to
validate the URL in the first run and then check if it already
exists in the second. This might not be the most efficient way
of handling it. The reason to do it that way is that it reduces
the amount of code that has to be replicated.