Links form order app to voucher/offer broken.
This is probably a breaking change for most people as the migration
history for the order app has been rewritten.  It will take some
fiddling around with south to get your system back in order.

This change is required though as we need to break the links between the
apps.