Adjust form_field.html to pivot on widget type
This requires a new templatetag that annotates the field with the widget
type.  The motivation is so that normal forms (not using
form-horizontal) can render checkboxes differently - wrapping the input
in the label element.  This is advocated by bootstrap amongst others:
http://twitter.github.com/bootstrap/base-css.html#forms

At the moment, checkboxes rendered in a form with class
'form-horizontal' don't look right - they need some adjusting in the
CSS.