Textareas in dashboard now have wysihtml widget
Almost all textareas have it enabled.  The corresponding variables in
the main templates have been marked 'safe'.

There are a few areas where the whitespace around the description isn't
quite right.  Will fix later.

The widget is enabled by setting a class on the form 'wysiwyg'.  It's
much easier to use the form classes as you don't normally have access to
the textarea classes directly.  To do that, you would need a custom form
widget and to build each form class to use it (rather than relying on
modelforms).  We may have to switch to this at some point if the current
implementation becomes too clunky.  We'll see.