Add reverse relations to preselect_related on category view
The availability wrapper needs to look at a product's class, which gives
two extra DB lookups for each product in the list.  This adds 50 DB
queries to the listing view and makes everybody sad.

This fix looks a little ugly but reduces the number of queries to ~25
(with cache enabled).