Injecting request object into AddToBasketForm
It was injecting user and basket separately before onto form contructor.
The result was that in some cases it would make it difficult to twist
functionality that was depending for instance on the request object.

Passing the whole request object on the form will make it more flexible
to change.

Also, the basket_tags where reviewed they also started passing the
request object instead of the user.