Abstract stock alert should be abstract
The `AbstractStockAlert` was missing `abstract=True` in its meta,
causing two concrete models to be created.  The subclasssing
`StockAlert` model was nothing more than a pointer to the original
fortunately meaning the migration back to a single concrete stock alert
model is not too painful.

The migration drops the stockalert table and renames the
partner_abstractstockalert table to partner_stockalert.  This should
work fine and not drop any data.

Thanks to @Tilley for finding this pesky oversight in #645.