Include 'back to search results' link on product detail page
This change introduces a templatetag which assigns context for adding a
'back to search results' link to a product detail page.  This can be
useful for certain types of user (who don't use the back button) and
want to get back to their, possibly complicated, set of search filters.

Fixes #456