Improve UX for deleted products
When a user ordered a product and it was deleted afterwards, there were
two problems with her order history view.
It wasn't working because the 'Add review' button relied on
line.product.slug, which threw a NoReverseMatch error.
The 'Reorder' button threw an error message '%title is no longer
available for purchase'. That was changed to '%title is no longer
available' to clarify.
Additionally, both the review and reorder buttons are now only displayed
when the product is still in our database - saving the user from a
disappointing click.