Automatically symlink "missing image" image into MEDIA_ROOT
The "missing image" is a static file and really shouldn't require that
hack of copying it to MEDIA_ROOT. This raises issues with versioning and
is an additional step in project setup.
This fix now tries to symlink the file from STATIC_ROOT to MEDIA_ROOT if
it doesn't exist.

Fixes #734