Introducing the "strategy" middleware
This is responsible for picking the appropriate strategy class for the
session and assigning it to the request.

The strategy class then is responsbile for returning the appropriate
price and availability wrappers for a given product.

A template tag is responsible for calling the strategy class with the
product to get the appropriate wrappers needed in the template.