Push translation files to Transifex
Currently, when commits to master add new strings, this has to be
manually pushed to Transifex - a process that traditionally only happens
before releases. This puts unnecessarily high pressure on translators to
translate the new strings in time for a release.

This is achieved with a little script and Travis' secret environment
variables protecting the account password of the Transifex account.