Use thumbnails for ImageInput widget
The entire image handling in the dashboard isn't very pretty. But at
least we're not using full-size images for this.
We also can't assume that just prepending MEDIA_URL is the correct URL
for every image (e.g. not when using S3 backends). Passing it through
Sorl gets us the correct URL for free.