Serve tracker pixel from Tangent's infrastructure
This is to enable some coarse metrics of who is using Oscar.  We just
look at the referer header to get a rough idea of which sites are using
Oscar.

This is an important measure for Tangent, Oscar's sponsor.  It's
documented in the release notes and we'll make sure everyone knows when
Oscar is released.  It's easy to opt out of.

Fixes #656