Prevent selecting non-shipping country as default address
In the address book form countries were not filtered by
is_shipping_country=True. And this is ok: for example the user has
a billing address not in a shipping country.

But one could set these addresses as default shipping address and they
appeared as valid shipping addresses. So users could make orders to
addresses in non-shipping countries.

@maikhoepfel squashed the commits, marked the error message as
translatable and simplified the template logic.

Reported by @itbabu.

Fixes #931