Re-bind image formset if other product formsets fail to validate
We currently don't hold on to images that are uploaded if any of the
other formsets fail when editing/creating products. To prevent
issue #1126, we need to re-bind the formset.

This behaviour should be improved as part of #856.

This commit was originally suggested in #1126,
re-done onto releases/0.6 in 05c185fc because of merge conflicts,
and now cherry-picked onto master again because it was easier again to
do so instead of fixing the merge conflicts.