Allow search facets to have options passed
I was testing this with adding a rating facet to the sandbox.  This
works ok but lots of products don't have a rating and so don't appear.
The syntax for getting Solr to select docs without a facet is geniunely
annoying and hard to get Haystack to do it.  I gave up.