Don't disable form button if HTML5 form is invalid
If HTML5 form has invalid form fields (like <input type="email">) the form will not be submitted on button click. The button must not be disabled to allow form submit after the form field is updated to be valid. Registration form is a good example, where if email is not valid it is not possible to submit the form after correcting input as the button stays disabled.

This change will not disable button on click if it is inside a form with invalid fields.