Don't special-case CommunicationEventType code in dashboard views
When sending out notifications, the code is looked up as-is. The
dashboard views previously first lower- and then upper-cased the code. I
fail to see a genuine reason for that, and dislike the added complexity.
More importantly, we don't enforce that the code is uppercased; one
can manually set the slug after all. This means a potential issue with
using slugs that become identical if both are lower-cased. E.g. both
'Foo' and 'FOO' are valid codes, but suddenly the lookup is ambiguous.