Make demo site in Travis
This ensures that we catch errors in the build process of the demo site,
and led to me spotting #1294.

I must've accidentally committed my demo database settings; switched to
a more generic defaults in demo/settings.py. Also added ATOMIC_REQUESTS
as that is recommended by Oscar.

django-extensions' reset_db fails with PostGIS without a template
database, hence it is skipped on Travis.
https://github.com/django-extensions/django-extensions/issues/489