Adjust AddToBasket form to require purchase info
This means we don't need to do the law-of-demeter-violating call to
fetch the purchase info for a product (line 79 of forms.py). HOWEVER,
I'm starting to have second thoughts about this - I've discovered that
the adding to basket process is quite confusing for variants and I
intend to refactor.

This commit also adds some useful tests though so it's a valid stepping
stone to a better place.