Simplify redirecting handling in some checkout views
Using the success_url attribute is common practice, makes it easier to
customise the URL and makes it more obvious that we only override
get_success_url() to attach a message.

Using the redirect shortcut instead of creating HttpResponseRedirects
also makes the code a bit cleaner.

This was brought to my attention because of #1343.

Also fixes the wrong class name for the bad request.