Management command to find duplicate email addresses
Oscar currently treats the local part of an email address in a
case-sensitive way. That's standards compliant, but the majority of
users is not aware of it and most SMTP server setups ignore casing.

In preparation for switching to a case-insensitive handling of email
addresses, this commit adds a management command to find duplicate email
addressses. Casing is ignored.

Related to #1179.