Smarter body handling for notifications
Notifications are only used for "Back in stock" notifications at the
moment; and those don't set a notification body, only a subject. The
template now shows a generic "Message" line if that's the case instead
of showing an empty body.

Furthermore, the notify_user[s] functions now accept arbitrary kwargs to
set on the Notification model, which allows setting custom fields, and
importantly, the notification body.

Fixes #1407.