Enhance tests/config.py to ease using the configured environment
It can sometimes be useful to be able to drop into the configured Django
environment of the test suite. It might e.g. be useful to drop into the
shell with ./config.py shell_plus or create missing migrations with
./config.py makemigrations

I needed this to create migrations for the test suite apps.