Fix deprecation path for renamed EmailBackend
Django stores the class name of the auth backend used to authenticate a
user. The simple assignment before meant that the new name was stored in
the session data, but it might not be in the available auth backends.
Hence logging in would fail as reported in #1432.

Inheriting from the new backend fixes the issue; logging in with the
deprecated name now works.

Fixes #1432.